Move type generator code out of util for cyclic import
